Implement navigateEvent.intercept() and navigateEvent.canIntercept
Follows https://github.com/WICG/navigation-api/pull/235

intercept() works very similarly to transitionWhile(), except that
insetad of taking a mandatory Promise, it takes an optional handler
function. If a function is provided and it returns a promise,
navigation finish will be delayed until the Promise resolve, just as
transitionWhile() delays navigation finish for its Promise.

canIntercept is identical to canTransition.

Bug: 1336000
Change-Id: I94edd7fdc727080594f16fe4511cb7c302d88941