Remove special FragmentData matching in NG fragmentation.
This code was needed by legacy fragmentation at some point, but not
anymore, and is only causing trouble for NG fragmentation, because it
would reset the clip rectangles (to "infinity"), causing us to hit
content that's clipped.

The root_layer->EnclosingPaginationLayer() == EnclosingPaginationLayer()
comparison was always true in NG, since EnclosingPaginationLayer() is
always nullptr then, which made us enter the special FragmentData
matching code as long as the object was fragmented at all.

Note that the new css-break/table/repeated-section/ tests were already
passing (with LayoutNGTableFragmentation) or failing (without it), but
we lacked test coverage.

Bug: 1330515
Change-Id: Ic47330264b58693e2b0c48019b22de10f00b7815